Testing moving the octokit mocking to its own module
